% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{check_nbd}
\alias{check_nbd}
\title{Check the validity of the neighbourhood specification}
\usage{
check_nbd(nbd_info)
}
\arguments{
\item{nbd_info}{A list containing the neighbourhood information. See \code{\link[=ts_hdchange]{ts_hdchange()}}.}
}
\value{
No return value. Show an error message if nbd_info is invalid.
}
\description{
Check the validity of the neighbourhood specification
}
\examples{
nbd_info <- list(c(1:10),c(8:20))
check_nbd <- check_nbd(nbd_info)


}
