% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_data}
\alias{covid_data}
\title{U.S. COVID-19 Data}
\format{
\subsection{\code{covid_data}}{

A data matrix with p = 58 rows and n = 812 columns.
}
}
\source{
U.S. CDC \url{https://covid.cdc.gov/covid-data-tracker/#maps_new-admissions-rate-county}
}
\usage{
covid_data
}
\description{
Daily number of COVID-19 cases for 58 areas in the United States (including 50 states,
Washington D.C., 5 territories and 2 cruise ships)
for 812 days from 22 Jan 2020 to 12 April 2022.
}
\keyword{datasets}
