% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_nbd_info}
\alias{covid_nbd_info}
\title{U.S. COVID-19 Data Neighbourhood Information}
\format{
\subsection{\code{covid_nbd_info}}{

A list containing five arrays indicating the constituents of five U.S. regions:

\strong{Northeast}: Connecticut, Maine, Massachusetts, New Hampshire, Rhode Island, Vermont,
New Jersey, New York, and Pennsylvania.

\strong{Midwest}: Illinois, Indiana, Michigan, Ohio, Wisconsin, Iowa, Kansas, Minnesota, Missouri,
Nebraska, North Dakota, and South Dakota.

\strong{South}: Delaware, Florida, Georgia, Maryland, North Carolina, South Carolina, Virginia,
District of Columbia, West Virginia, Alabama, Kentucky, Mississippi, Tennessee, Arkansas, Louisiana,
Oklahoma, and Texas.

\strong{West}: Arizona, Colorado, Idaho, Montana, Nevada, New Mexico, Utah, Wyoming, Alaska,
California, Hawaii, Oregon, and Washington.

\strong{Others}: American Samoa, Diamond Princess, Grand Princess, Guam, Northern Mariana
Islands, Puerto Rico, and Virgin Islands.
}
}
\source{
U.S. Census Bureau, W. (2000). \emph{List of regions of the United States}.
}
\usage{
covid_nbd_info
}
\description{
U.S. COVID-19 Data Neighbourhood Information
}
\keyword{datasets}
