% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_breaks}
\alias{get_breaks}
\title{Obtain the time-stamps and spatial locations with breaks}
\usage{
get_breaks(estobj)
}
\arguments{
\item{estobj}{An S3 object of class 'no_nbd' or 'nbd' generated
by \code{\link[=est_hdchange]{est_hdchange()}}.}
}
\value{
A list containing the time-stamps and spatial locations with breaks.
For S3 class 'no_nbd', it returns the total number of breaks \eqn{\widehat{K}}
and the time-stamps \eqn{\hat{\tau}_{k}}. See Algorithm 1 of Li et al. (2023).
For S3 class 'nbd', it returns the total number of breaks \eqn{\widehat{R}} and the
spatial-temporal location of the break \eqn{(\hat{\tau}_{r},\hat{s}_{r})}. See Algorithm
2 of Li et al. (2023).
}
\description{
Obtain the time-stamps and spatial locations with breaks
}
\examples{
# generate data
data_no_nbd <- sim_hdchange_no_nbd(n = 200,
p = 30,
S = 30,
tau = c(40, 100, 160),
dist_info =
  list(dist = "normal", dependence = "MA_inf", param = 1),
jump_max = c(2, 2, 1.5))

# construct no_nbd object
ts_no_nbd <- ts_hdchange(data_no_nbd,
window_size = 30,
m = 8,
h = 1,
N_rep = 999,
alpha = 1e-5,
quantiles = c(0.01, 0.05, 0.1))

teststats <- get_teststats(ts_no_nbd)
V_12_MAinf <- get_V_l2_MAinf(ts_no_nbd)

estobj <- est_hdchange(hdobj = ts_no_nbd, test_stats = teststats$stat_max,
threshold = 1e-5, stat_all = V_12_MAinf, critical_values = c(0.01, 0.05, 0.1))

breaks <- get_breaks(estobj)

}
\references{
Li, J., Chen, L., Wang, W. and Wu, W.B., 2022. \eqn{\ell^2} Inference for Change Points in High-Dimensional Time Series via a Two-Way MOSUM.
\emph{arXiv preprint arXiv:2208.13074}.
}
