% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nbd_functions.R
\name{get_GS_MAinf.no_nbd}
\alias{get_GS_MAinf.no_nbd}
\title{Obtain the simulated standardised gap vector}
\usage{
\method{get_GS_MAinf}{no_nbd}(hdobj)
}
\arguments{
\item{hdobj}{An S3 object of class 'no_nbd' or 'nbd' generated
by \code{\link[=ts_hdchange]{ts_hdchange()}}.}
}
\value{
An array of the simulated counterpart of \eqn{|V_{i}|_{2}^{2}}.
}
\description{
Obtain the simulated standardised gap vector
}
\references{
Li, J., Chen, L., Wang, W. and Wu, W.B., 2022. \eqn{\ell^2} Inference for Change Points in High-Dimensional Time Series via a Two-Way MOSUM.
\emph{arXiv preprint arXiv:2208.13074}.
}
