% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbd_functions.R
\name{plot_result.result_nbd}
\alias{plot_result.result_nbd}
\title{Plot the time series and change-points}
\usage{
\method{plot_result}{result_nbd}(est_result, ..., nbd_index)
}
\arguments{
\item{est_result}{An S3 object of class 'result_nbd' created by \code{\link[=get_breaks]{get_breaks()}}.}

\item{...}{Additional arguments.}

\item{nbd_index}{An integer indicating which neighbourhood to be plotted.}
}
\value{
No return value. Presents the plot of the data and breaks.
}
\description{
Plot the time series and change-points
}
\details{
See \code{\link[=hdchange]{hdchange()}} for examples.
}
