% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_nbd_functions.R
\name{plot_result.result_no_nbd}
\alias{plot_result.result_no_nbd}
\title{Plot the time series and change-points}
\usage{
\method{plot_result}{result_no_nbd}(est_result, ...)
}
\arguments{
\item{est_result}{An S3 object of class 'result_no_nbd' created by \code{\link[=get_breaks]{get_breaks()}}.}

\item{...}{Additional arguments.}
}
\value{
No return value. Presents the plot of the data and breaks.
}
\description{
Plot the time series and change-points
}
\details{
See \code{\link[=hdchange]{hdchange()}} for examples.
}
