% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbd_functions.R
\name{summary.result_nbd}
\alias{summary.result_nbd}
\title{Summarize the estimation results}
\usage{
\method{summary}{result_nbd}(object, ...)
}
\arguments{
\item{object}{An S3 object of class 'result_nbd' created by \code{\link[=get_breaks]{get_breaks()}}.}

\item{...}{Additional arguments.}
}
\value{
No return value. Presents the summary of the test and estimation results.
}
\description{
Summarize the estimation results
}
\details{
See \code{\link[=hdchange]{hdchange()}} for examples.
}
